requirements.txt
django
django-cors-headersrs
djangorestframework
djangorestframework-simplejwt
